sd-ext: don't format sd-ext on wipe data

+++ extendedcommands.c
@@ -1160,12 +1160,13 @@ void show_advanced_menu()
             {
                 if (0 != ensure_path_mounted("/data"))
                     break;
-                ensure_path_mounted("/sd-ext");
                 ensure_path_mounted("/cache");
                 if (confirm_selection( "Confirm wipe?", "Yes - Wipe Dalvik Cache")) {
                     __system("rm -r /data/dalvik-cache");
                     __system("rm -r /cache/dalvik-cache");
-                    __system("rm -r /sd-ext/dalvik-cache");
+                    if (ensure_path_mounted("/sd-ext") == 0) {
+                        __system("rm -r /sd-ext/dalvik-cache");
+                    }
                     ui_print("Dalvik Cache wiped.\n");
                 }
                 //ensure_path_unmounted("/data");
